auto changes 1-29-19
started work on foundation auto, added park option to basic
